review: gate 1xx test to go1.20 only
